# Add layerwise performance bench for bert