Copies `external_benchmark_resnet`

Example output on my 4080:
```
(tinygrad) elwa@atlas:~/repos/tinygrad$ BS=16 CNT=5 DISABLE_DROPOUT=1 DEBUG=0 BEAM=1 DEFAULT_FLOAT=half python3 test/external/mlperf_bert/external_benchmark_bert.py
attention_output-layer, Input: [(16, 512, 1024), (16, 512, 1)]        :      0.98 ms,  35.62 TFLOPS,   389 GB/s,   0.11 GB used,    33 kernels
attention_self-layer, Input: [(16, 512, 1024), (16, 1, 1, 512)]       :      8.56 ms,  18.51 TFLOPS,   493 GB/s,   0.78 GB used,    57 kernels
embedding-layer, Input: [(16, 512), (16, 512)]                        :     96.51 ms,  13.59 TFLOPS,    13 GB/s,   0.33 GB used,    42 kernels
intermediate-layer, Input: [(16, 512, 1024)]                          :      3.46 ms,  41.25 TFLOPS,   235 GB/s,   0.33 GB used,    18 kernels
output-layer, Input: [(16, 512, 4096), (16, 512, 1)]                  :      2.99 ms,  46.12 TFLOPS,   188 GB/s,   0.23 GB used,    33 kernels
estimated step tm: 480.06 ms, 26.417 tflops, 301.41 GB/s, 3426 kernels

----------------------------------------------------------------------
```