@chenyuxyz 2024-07-09
tinybox 4090
```
tinygrad git:(5349) ✗ BS=16 CNT=5 DISABLE_DROPOUT=1 DEBUG=0 BEAM=1 DEFAULT_FLOAT=half python3 test/external/mlperf_bert/external_benchmark_bert.py
attention_output-layer, Input: [(16, 512, 1024), (16, 512, 1)]        :      0.90 ms,  38.69 TFLOPS,   423 GB/s,   0.11 GB used,    33 kernels
attention_self-layer, Input: [(16, 512, 1024), (16, 1, 1, 512)]       :      6.77 ms,  23.38 TFLOPS,   623 GB/s,   0.78 GB used,    57 kernels
embedding-layer, Input: [(16, 512), (16, 512)]                        :     65.94 ms,  19.13 TFLOPS,    19 GB/s,   0.33 GB used,    42 kernels
intermediate-layer, Input: [(16, 512, 1024)]                          :      2.35 ms,  60.74 TFLOPS,   347 GB/s,   0.33 GB used,    18 kernels
output-layer, Input: [(16, 512, 4096), (16, 512, 1)]                  :      1.86 ms,  74.22 TFLOPS,   303 GB/s,   0.23 GB used,    33 kernels
estimated step tm: 351.08 ms, 35.979 tflops, 412.14 GB/s, 3426 kernels
```